+++ Changes.md
@@ -15,6 +15,7 @@
   sibling `Array` methods. Having an Parse-JSON Hash array or a Parse::CollectionProxy which contains a series
   of Parse hashes can now be easily converted to an array of Parse objects with these methods.
 - FIXED: Correctly discards ACL changes on User model saves.
+- FIXED: Fixes issues with double '/' in update URI paths.
 
 ### 1.7.3
 - CHANGED: Moved to using preferred ENV variable names based on parse-server cli.